reduce size of large test file
